third_party/blink/web_tests/platform/generic/plugins/plugin-paint-test-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
This tests that dynamically added plugins receive paint events on creation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS plugin message: loaded
PASS successfullyParsed is true

TEST COMPLETE
